Move threadpool namespace
